requirements.txt
openai>=1.88.0
python-dotenv>=1.0.0
reportlab>=3.6.0
requests>=2.31.0